1- Added Integration Tests for Inventory 2- fix issues that appeared after added these tests

+++ commercetools.Sdk/commercetools.Sdk.Domain/InventoryEntry/InventoryEntryDraft.cs
@@ -8,7 +8,7 @@
         [Required]
         public string Sku { get; set; }
         public Reference<Channel> SupplyChannel { get; set; }
-        public double QuantityOnStock { get; set; }
+        public long QuantityOnStock { get; set; }
         public int RestockableInDays { get; set; }
         public DateTime ExpectedDelivery { get; set; }
         public CustomFields Custom { get; set; }
